added simulate
